refactor(worker): restructure src into infrastructure/ + strategies/
