+++ packages/worker/src/deployments/processor.ts
@@ -1,9 +1,12 @@
 import { db } from "../config/db.js";
 import { getEnv } from "../config/env.js";
 import { logger } from "../config/logger.js";
-import { upsertFileRoute, upsertProxyRoute } from "./caddy/client.js";
-import { DockerRunner } from "./docker/docker-runner.js";
-import { DeploymentOrchestrator } from "./orchestrator.js";
+import {
+	upsertFileRoute,
+	upsertProxyRoute,
+} from "../infrastructure/caddy/client.js";
+import { DockerRunner } from "../infrastructure/docker/docker-runner.js";
+import { DeploymentOrchestrator } from "./pipeline.js";
 
 export async function processDeployment(deploymentId: string): Promise<void> {
 	const orchestrator = new DeploymentOrchestrator({
