--- packages/worker/src/jobs/deploy.ts
@@ -1,1 +1,0 @@
-export { processDeployment } from "../deployments/process-deployment.js";
